did not find any duplicate components, but added an editor script to also do this automatically
